cloud/functions/generate_prompt/requirements.txt
functions-framework==3.*
firebase-admin==6.*
google-generativeai==0.8.*
